+++ scripts/deploy/contracts/artifacts/exportArtifacts.js
@@ -1,8 +1,6 @@
 /* eslint-disable no-console */
-const writeArtifact = require('./writeArtifact')
-
-const zosGetProxyAddress = require('../zos/contracts/addresses/getProxyAddress')
-const zosGetMigrations = require('../zos/handlers/getMigrations')
+const exportContractArtifacts = require('./exportContractArtifacts')
+const exportLibraryArtifacts = require('./exportLibraryArtifacts')
 
 async function exportArtifacts(
     projectName,
@@ -15,52 +13,21 @@ async function exportArtifacts(
         console.log(`Exporting: ${projectName}`)
     }
 
-    // load migrations from zos
-    const { contracts, solidityLibs } = zosGetMigrations(networkId)
-    const contractNames = Object.keys(contracts)
+    await exportContractArtifacts(
+        projectName,
+        networkName,
+        networkId,
+        version,
+        verbose
+    )
 
-    for (const contractName of contractNames) {
-        if (verbose) {
-            console.log(`Exporting artifact: ${contractName}.${networkName}.json`)
-        }
-
-        // get proxy address from zos proxies
-        const proxyAddress = await zosGetProxyAddress(
-            projectName,
-            contractName,
-            networkId
-        )
-
-        const artifact = writeArtifact(
-            contractName,
-            proxyAddress,
-            networkName,
-            version
-        )
-
-        if (verbose) {
-            console.log(`Exported contract artifact: ${artifact.version} of ${contractName} at ${artifact.address}`)
-        }
-    }
-
-    const solidityLibNames = Object.keys(solidityLibs)
-
-    for (const solidityLibName of solidityLibNames) {
-        if (verbose) {
-            console.log(`Exporting library: ${solidityLibName}.${networkName}.json`)
-        }
-
-        const artifact = writeArtifact(
-            solidityLibName,
-            solidityLibs[solidityLibName].address,
-            networkName,
-            version
-        )
-
-        if (verbose) {
-            console.log(`Exported library artifact: ${artifact.version} of ${solidityLibName} at ${artifact.address}`)
-        }
-    }
+    await exportLibraryArtifacts(
+        projectName,
+        networkName,
+        networkId,
+        version,
+        verbose
+    )
 
     if (verbose) {
         console.log(projectName, version, networkName)
